Always generate SSL certificates

+++ src/wok/proxy.py
@@ -33,10 +33,6 @@ DH_COMMAND = 'openssl dhparam -dsaparam -out %s 2048'
 
 
 def check_proxy_config():
-    # When running from a installed system, there is nothing to do
-    if paths.installed:
-        return
-
     # Otherwise, ensure essential directories and files are placed on right
     # place to avoid problems
     #
@@ -51,11 +47,12 @@ def check_proxy_config():
     # running from source code
     symlinks = [{'target': os.path.join(paths.nginx_conf_dir, 'wok.conf'),
                  'link': os.path.join(paths.sys_nginx_conf_dir, 'wok.conf')}]
-    for item in symlinks:
-        link = item['link']
-        if os.path.isfile(link) or os.path.islink(link):
-            os.remove(link)
-        os.symlink(item['target'], link)
+    if not paths.installed:
+        for item in symlinks:
+            link = item['link']
+            if os.path.isfile(link) or os.path.islink(link):
+                os.remove(link)
+            os.symlink(item['target'], link)
 
     # Generate unique Diffie-Hellman group with 2048-bit
     dh_file = os.path.join(paths.sys_conf_dir, 'dhparams.pem')
